move pipelines tests into `pipelines` subdirectory to mirror command structure
