+++ tests/pipelines/test_bump_version.py
