+++ tests/pipelines/test_create_logo.py
